+++ tests/infer/test_contract.py
@@ -1,0 +1,50 @@
+from __future__ import absolute_import, division, print_function
+
+import pytest
+import torch
+
+from pyro.infer.contract import _partition_terms
+
+
+@pytest.mark.parametrize('shapes,dims,expected_num_components', [
+    ([()], set(), 1),
+    ([(2,)], set(), 1),
+    ([(2,)], set([-1]), 1),
+    ([(2,), (2,)], set(), 2),
+    ([(2,), (2,)], set([-1]), 1),
+    ([(2, 1), (2, 1), (1, 3), (1, 3)], set(), 4),
+    ([(2, 1), (2, 1), (1, 3), (1, 3)], set([-1]), 3),
+    ([(2, 1), (2, 1), (1, 3), (1, 3)], set([-2]), 3),
+    ([(2, 1), (2, 1), (1, 3), (1, 3)], set([-1, -2]), 2),
+    ([(2, 1), (2, 3), (1, 3)], set(), 3),
+    ([(2, 1), (2, 3), (1, 3)], set([-1]), 2),
+    ([(2, 1), (2, 3), (1, 3)], set([-2]), 2),
+    ([(2, 1), (2, 3), (1, 3)], set([-1, -2]), 1),
+    ([(4, 1, 1), (4, 3, 1), (1, 3, 2), (1, 1, 2)], set(), 4),
+    ([(4, 1, 1), (4, 3, 1), (1, 3, 2), (1, 1, 2)], set([-1]), 3),
+    ([(4, 1, 1), (4, 3, 1), (1, 3, 2), (1, 1, 2)], set([-2]), 3),
+    ([(4, 1, 1), (4, 3, 1), (1, 3, 2), (1, 1, 2)], set([-3]), 3),
+    ([(4, 1, 1), (4, 3, 1), (1, 3, 2), (1, 1, 2)], set([-1, -3]), 2),
+    ([(4, 1, 1), (4, 3, 1), (1, 3, 2), (1, 1, 2)], set([-1, -2, -3]), 1),
+])
+def test_partition_terms(shapes, dims, expected_num_components):
+    tensors = [torch.randn(shape) for shape in shapes]
+    components = list(_partition_terms(tensors, dims))
+
+    # Check that result is a partition.
+    expected_terms = sorted(tensors, key=id)
+    actual_terms = sorted((x for c in components for x in c[0]), key=id)
+    assert actual_terms == expected_terms
+    assert dims == set.union(set(), *(c[1] for c in components))
+
+    # Check that partition is not too coarse.
+    assert len(components) == expected_num_components
+
+    # Check that partition is not too fine.
+    component_dict = {x: i for i, (terms, _) in enumerate(components) for x in terms}
+    for x in tensors:
+        for y in tensors:
+            if x is not y:
+                if any(dim >= -x.dim() and x.shape[dim] > 1 and
+                       dim >= -y.dim() and y.shape[dim] > 1 for dim in dims):
+                    assert component_dict[x] == component_dict[y]
